\name{JDL-package}
\alias{JDL-package}
\alias{JDL}
\docType{package}
\title{
\packageTitle{JDL}
}
\description{
\packageDescription{JDL}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{JDL}
\packageIndices{JDL}
}
\author{
\packageAuthor{JDL}

Maintainer: \packageMaintainer{JDL}
}
\references{
~~ Literature or other references for background information ~~
}

\keyword{ 
	~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
	package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
